Moved Add forms to modal
