+++ package.json
@@ -23,6 +23,7 @@
     "next-auth": "^4.23.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-icons": "^4.11.0",
     "superjson": "^1.13.1",
     "zod": "^3.21.4"
   },
